requirements.txt
annotated-types==0.7.0
black==24.4.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
confuse==2.0.1
idna==3.7
ijson==3.3.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-amazon-sp-api==1.5.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
